requirements.txt
apache-airflow==3.0.2
pandas==2.1.4
SQLAlchemy==1.4.54
